requirements.txt
beautifulsoup4 == 4.9.3
biopython == 1.78
gensim == 3.8.1
networkx == 2.5
numpy == 1.19.2
pandas == 1.1.3
pymongo == 3.11.3
pytextrank == 3.1.1
python_docx == 0.8.10
pytrials == 0.1.2
requests == 2.24.0
scikit_learn == 0.23.2
selenium == 3.141.0
spacy == 3.0.6
streamlit == 0.80.0
summarizer == 0.0.7
torch == 1.8.1
transformers == 4.5.1
